backend/requirements.txt
arabic-reshaper==3.0.0
asgiref==3.9.1
asn1crypto==1.5.1
certifi==2025.8.3
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.3
cryptography==45.0.7
cssselect2==0.8.0
Django==5.2.5
django-cors-headers==4.7.0
djangorestframework==3.16.1
et_xmlfile==2.0.0
html5lib==1.1
idna==3.10
lxml==6.0.1
mysqlclient==2.2.7
numpy==2.3.3
openpyxl==3.1.5
oscrypto==1.3.0
pandas==2.3.2
pillow==11.3.0
pycparser==2.23
pyHanko==0.30.0
pyhanko-certvalidator==0.28.0
pypdf==6.0.0
pypng==0.20220715.0
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.2
qrcode==7.4.2
reportlab==4.2.0
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
svglib==1.5.1
tinycss2==1.4.0
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.5.0
webencodings==0.5.1
xhtml2pdf==0.2.17

xlwt==1.3.0
